Implement uniq visits functionality.
At this point the whole functionality of the test is beeing implemented.
However there is quite a bit of code duplication at the moment so we are
going to try to refactor the code to remove duplitions as much as possible.

Integration test has been added for both total visits and uniq visits. At this
point the is 100% coverage which will now help to refactor the code.